--- a/cmd/gitserver/server/server_test.go
+++ b/cmd/gitserver/server/server_test.go
@@ -932,7 +932,6 @@ var ignoreVolatileGitserverRepoFields = cmpopts.IgnoreFields(
 	"UpdatedAt",
 	"CorruptionLogs",
 	"CloningProgress",
-	"LastSyncOutput",
 )
 
 func TestHandleRepoDelete(t *testing.T) {