We recently added a left join on another table holding the gitserver sync logs, which added quite some overhead to the queries. Here's an example query plan from dotcom for the iterate gitserver repos query:

Before

```
SELECT
    gr.repo_id, repo.name, gr.clone_status, gr.cloning_progress, gr.shard_id, gr.last_error, gr.last_fetched, gr.last_changed, gr.repo_size_bytes, gr.updated_at, gr.corrupted_at, gr.corruption_logs, gr.pool_repo_id, go.last_output 
FROM gitserver_repos gr 
JOIN repo ON gr.repo_id = repo.id 
LEFT OUTER JOIN gitserver_repos_sync_output go ON gr.repo_id = go.repo_id 
WHERE gr.repo_id > 0 AND repo.id > 0 
ORDER BY gr.repo_id ASC 
LIMIT 5000 
OFFSET 0;
```

```
                                                                                           QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=4.83..1408.91 rows=5000 width=489) (actual time=0.155..2133.020 rows=5000 loops=1)
   ->  Merge Left Join  (cost=4.83..2619877.12 rows=9329510 width=489) (actual time=0.154..2132.011 rows=5000 loops=1)
         Merge Cond: (gr.repo_id = go.repo_id)
         ->  Merge Join  (cost=4.40..2219410.84 rows=9329510 width=328) (actual time=0.024..47.463 rows=5000 loops=1)
               Merge Cond: (gr.repo_id = repo.id)
               ->  Index Scan using gitserver_repos_pkey on gitserver_repos gr  (cost=0.56..884164.75 rows=9329510 width=285) (actual time=0.008..24.785 rows=5000 loops=1)
                     Index Cond: (repo_id > 0)
               ->  Index Scan using repo_pkey on repo  (cost=0.43..1197363.74 rows=9399073 width=47) (actual time=0.009..16.556 rows=5000 loops=1)
                     Index Cond: (id > 0)
         ->  Index Scan using gitserver_repos_sync_output_pkey on gitserver_repos_sync_output go  (cost=0.43..306938.95 rows=5619062 width=165) (actual time=0.129..2076.751 rows=4738 loops=1)
 Planning Time: 3.811 ms
 Execution Time: 2134.226 ms
(12 rows)
```

After

```
EXPLAIN ANALYZE
SELECT
    gr.repo_id, repo.name, gr.clone_status, gr.cloning_progress, gr.shard_id, gr.last_error, gr.last_fetched, gr.last_changed, gr.repo_size_bytes, gr.updated_at, gr.corrupted_at, gr.corruption_logs, gr.pool_repo_id
FROM gitserver_repos gr 
JOIN repo ON gr.repo_id = repo.id 
WHERE gr.repo_id > 0 AND repo.id > 0 
ORDER BY gr.repo_id ASC 
LIMIT 5000 
OFFSET 0;
```

```
                                                                              QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=4.40..1193.85 rows=5000 width=328) (actual time=0.023..34.821 rows=5000 loops=1)
   ->  Merge Join  (cost=4.40..2219410.84 rows=9329510 width=328) (actual time=0.022..34.156 rows=5000 loops=1)
         Merge Cond: (gr.repo_id = repo.id)
         ->  Index Scan using gitserver_repos_pkey on gitserver_repos gr  (cost=0.56..884164.75 rows=9329510 width=285) (actual time=0.008..20.225 rows=5000 loops=1)
               Index Cond: (repo_id > 0)
         ->  Index Scan using repo_pkey on repo  (cost=0.43..1197363.74 rows=9399073 width=47) (actual time=0.011..10.150 rows=5000 loops=1)
               Index Cond: (id > 0)
 Planning Time: 0.732 ms
 Execution Time: 35.279 ms
(9 rows)
```

Since we only used this field in a single place, I decided to create a separate query function for it, in favor of dealing with complexity of conditionally loading or not loading this field.

## Test plan

Added tests, query plans above.